Implement workaround for util.download_gf_family.
Previous implementation would 'open' a font from a zipfile. Opened
files from zipfile objects are unseekable streams. FontTools recently
lost support to intantiate a TTFont object from such streams,
https://github.com/fonttools/fonttools/issues/1362

New implementation will 'read' the font from a zipfile which returns
the bytes of the read file. The bytes are then stored as a StringIO
object which is a seekable stream.